pyproject.toml
[project]
name = "slackbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.28.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.13.5",
    "openai>=2.32.0",
    "rapidfuzz>=3.14.5",
    "google-api-python-client>=2.170.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
